+++ src/codegen/operator/global_group_by_translator.cpp
@@ -12,7 +12,7 @@
 
 #include "codegen/operator/global_group_by_translator.h"
 
-#include "include/codegen/util/if.h"
+#include "codegen/util/if.h"
 #include "common/logger.h"
 
 namespace peloton {
