+++ src/codegen/operator/table_scan_translator.cpp
@@ -12,9 +12,9 @@
 
 #include "codegen/operator/table_scan_translator.h"
 
-#include "include/codegen/util/if.h"
-#include "include/codegen/proxy/catalog_proxy.h"
-#include "include/codegen/proxy/transaction_runtime_proxy.h"
+#include "codegen/util/if.h"
+#include "codegen/proxy/catalog_proxy.h"
+#include "codegen/proxy/transaction_runtime_proxy.h"
 #include "planner/seq_scan_plan.h"
 #include "storage/data_table.h"
