+++ src/codegen/sorter.cpp
@@ -12,8 +12,8 @@
 
 #include "codegen/sorter.h"
 
-#include "include/codegen/util/loop.h"
-#include "include/codegen/proxy/sorter_proxy.h"
+#include "codegen/util/loop.h"
+#include "codegen/proxy/sorter_proxy.h"
 #include "codegen/vectorized_loop.h"
 #include "codegen/vector.h"
