+++ src/codegen/tile_group.cpp
@@ -13,11 +13,11 @@
 #include "codegen/tile_group.h"
 
 #include "catalog/schema.h"
-#include "include/codegen/util/if.h"
-#include "include/codegen/util/loop.h"
-#include "include/codegen/proxy/runtime_functions_proxy.h"
+#include "codegen/util/if.h"
+#include "codegen/util/loop.h"
+#include "codegen/proxy/runtime_functions_proxy.h"
+#include "codegen/proxy/tile_group_proxy.h"
 #include "codegen/scan_consumer.h"
-#include "include/codegen/proxy/tile_group_proxy.h"
 #include "codegen/type.h"
 #include "codegen/varlen.h"
 #include "codegen/vector.h"
