+++ src/codegen/utils/loop.cpp
@@ -10,11 +10,10 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "include/codegen/util/loop.h"
+#include "include/codegen/utils/loop.h"
 
 namespace peloton {
 namespace codegen {
-namespace util {
 
 // Constructor
 Loop::Loop(CodeGen &cg, llvm::Value *start_condition,
@@ -76,6 +75,5 @@ void Loop::CollectFinalLoopVariables(std::vector<llvm::Value *> &loop_vals) {
   }
 }
 
-}  // namespace util
 }  // namespace codegen
 }  // namespace peloton
