+++ src/include/codegen/utils/loop.h
@@ -18,7 +18,6 @@
 
 namespace peloton {
 namespace codegen {
-namespace util {
 
 //===----------------------------------------------------------------------===//
 // A utility class to help generate loops in LLVM IR
@@ -63,6 +62,5 @@ class Loop {
   std::vector<llvm::PHINode *> phi_nodes_;
 };
 
-}  // namespace util
 }  // namespace codegen
 }  // namespace peloton
